fix:fix the type of carbon from Long to Double

+++ src/main/java/com/example/EcoGo/dto/FacultyStatsDto.java
@@ -14,9 +14,9 @@ public class FacultyStatsDto {
 
     public static class CarbonResponse {
         public String faculty;
-        public Long totalCarbon;
+        public Double totalCarbon;
 
-        public CarbonResponse(String faculty, Long totalCarbon) {
+        public CarbonResponse(String faculty, Double totalCarbon) {
             this.faculty = faculty;
             this.totalCarbon = totalCarbon;
         }
